fix: update SDK test files to use refactored API
- Remove PAPITransactionSubmitter usage (removed in refactor)
- Update import paths for new typescript/ directory structure
- Use AsyncBulletinClient directly with PAPI client and signer
- Update paths: ./common.js -> ./typescript/common.js
- Update paths: ./.papi/... -> ./typescript/.papi/...
